feat(Base): add empty static innerTypes
fixes #263
